requirements.txt
aiolimiter==1.1.0
anyio==4.6.2.post1
APScheduler==3.6.3
beautifulsoup4==4.12.2
cachetools==4.2.2
certifi==2023.11.17
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.1
et-xmlfile==1.1.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.6
httpx==0.27.2
hyperframe==6.0.1
idna==3.6
numpy==1.26.2
openpyxl==3.1.2
pandas==2.1.3
pycparser==2.22
pyTelegramBotAPI==4.23.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
requests==2.31.0
setuptools==75.2.0
six==1.16.0
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.5
tornado==6.4.1
tzdata==2023.3
tzlocal==5.2
urllib3==2.1.0
